+++ src/main/java/com/github/toolarium/processing/unit/runtime/runnable/IProcessingUnitRunnableListener.java
@@ -5,10 +5,14 @@
  */
 package com.github.toolarium.processing.unit.runtime.runnable;
 
+import com.github.toolarium.common.bandwidth.IBandwidthThrottling;
 import com.github.toolarium.processing.unit.IProcessingUnitContext;
 import com.github.toolarium.processing.unit.IProcessingUnitProgress;
+import com.github.toolarium.processing.unit.dto.Parameter;
 import com.github.toolarium.processing.unit.dto.ProcessingActionStatus;
+import com.github.toolarium.processing.unit.dto.ProcessingRuntimeStatus;
 import com.github.toolarium.processing.unit.runtime.IProcessingUnitRuntimeTimeMeasurement;
+import java.util.List;
 
 
 /**
@@ -18,8 +22,41 @@ import com.github.toolarium.processing.unit.runtime.IProcessingUnitRuntimeTimeMe
  */
 public interface IProcessingUnitRunnableListener {
     
+
     /**
-     * Notify processing unit action status
+     * Notify processing unit progress. It is called if at least 1 percent progress has been made. 
+     * The processingProgress.getProgress() contains the progress, the last progress in percentage 
+     * is past as well. 
+     *
+     * @param id the id
+     * @param name the name
+     * @param processingUnitClass the processing unit class
+     * @param parameters the parameters
+     * @param processingUnitContext the processing unit context
+     * @param processingProgress the progressing unit progress
+     * @param processingActionStatus the action status
+     * @param processingRuntimeStatus the runtime status
+     * @param messages the messages
+     * @param timeMeasurement the time measurement
+     * @param processingUnitThrottling the processing unit throttling
+     * @param lastProgressInPercentage the last progress in percentage
+     */
+    void notifyProcessingUnitProgress(String id, // CHECKSTYLE IGNORE THIS LINE
+                                      String name, 
+                                      String processingUnitClass,
+                                      List<Parameter> parameters,
+                                      IProcessingUnitContext processingUnitContext,
+                                      IProcessingUnitProgress processingProgress, 
+                                      ProcessingActionStatus processingActionStatus,
+                                      ProcessingRuntimeStatus processingRuntimeStatus,
+                                      List<String> messages,
+                                      IProcessingUnitRuntimeTimeMeasurement timeMeasurement, 
+                                      IBandwidthThrottling processingUnitThrottling,
+                                      int lastProgressInPercentage);
+    
+    
+    /**
+     * Notification processing unit an action status has changed.
      *
      * @param id the unique id of this processing 
      * @param name the name of this processing unit runnable